backend/requirement.txt
flask
flask-cors
pymongo[srv]
web3
eth-account
python-dotenv